+++ benchmarks/Bench.hs
@@ -19,7 +19,7 @@ import           Control.DeepSeq (NFData)
 import           Criterion.Main (Benchmark, bench, bgroup, defaultMain, nf)
 
 import           Data.List (foldl')
-import qualified Data.Text.Lazy as T (pack)
+import qualified Data.Text as T
 
 import           GHC.Generics (Generic)
 
@@ -31,9 +31,20 @@ main :: IO ()
 main = defaultMain
     [ sampleGroup "String Show"                 BTLeaf1 BTBranch1 BTEmpty1 show
     , sampleGroup "String Show, then Text.pack" BTLeaf1 BTBranch1 BTEmpty1 (T.pack . show)
-    , sampleGroup "Text Show (TH)"              BTLeaf2 BTBranch2 BTEmpty2 showtl
-    , sampleGroup "Text Show (generics)"        BTLeaf3 BTBranch3 BTEmpty3 showtl
+    , sampleGroup "TextShow (TH)"               BTLeaf2 BTBranch2 BTEmpty2 showt
+    , sampleGroup "TextShow (generics)"         BTLeaf3 BTBranch3 BTEmpty3 showt
+    , bgroup "Enumeration type"
+      [ bench "String Show"                 $ nf show            Violet
+      , bench "String Show, then Text.pack" $ nf (T.pack . show) Violet
+      , bench "TextShow (TH)"               $ nf showt           Violet
+      , bench "TextShow (generics)"         $ nf showt         $ Color2 Violet
+      , bench "Manually written showt"      $ nf colorShowt      Violet
+      ]
     ]
+
+-------------------------------------------------------------------------------
+-- Tree-like ADTs
+-------------------------------------------------------------------------------
 
 sampleGroup :: forall a b. NFData b
             => String -> (Int -> a) -> (a -> a -> a) -> a -> (a -> b) -> Benchmark
@@ -56,27 +67,24 @@ type Sample = forall a b.
 
 smallSample :: Sample
 smallSample (leaf, branch, _, showFun) =
-    showFun $
-        (leaf 12345 `branch` leaf 1234) `branch`
-        leaf 123456 `branch`
-        (leaf 1234567 `branch` leaf 123456)
+    showFun $ sampleTree leaf branch
 {-# NOINLINE smallSample #-}
 
 mediumSample :: Sample
 mediumSample (leaf, branch, empty, showFun) =
-    showFun . foldl' branch empty . replicate 1000 $
-        (leaf 12345 `branch` leaf 1234) `branch`
-        leaf 123456 `branch`
-        (leaf 1234567 `branch` leaf 123456)
+    showFun . foldl' branch empty . replicate 1000 $ sampleTree leaf branch
 {-# NOINLINE mediumSample #-}
 
 largeSample :: Sample
 largeSample (leaf, branch, empty, showFun) =
-    showFun . foldl' branch empty . replicate 100000 $
-        (leaf 12345 `branch` leaf 1234) `branch`
-        leaf 123456 `branch`
-        (leaf 1234567 `branch` leaf 123456)
+    showFun . foldl' branch empty . replicate 100000 $ sampleTree leaf branch
 {-# NOINLINE largeSample #-}
+
+sampleTree :: (Int -> a) -> (a -> a -> a) -> a
+sampleTree leaf branch =
+    (leaf 12345 `branch` leaf 1234) `branch`
+    leaf 123456 `branch`
+    (leaf 1234567 `branch` leaf 123456)
 
 -- NB: constructors must be same length!
 data BinTree1 a = BTEmpty1
@@ -96,4 +104,27 @@ data BinTree3 a = BTEmpty3
 instance TextShow a => TextShow (BinTree3 a) where
     showbPrec = genericShowbPrec
 
+-------------------------------------------------------------------------------
+-- Simple enumeration types
+-------------------------------------------------------------------------------
+
+data Color = Red | Green | Blue | Orange | Violet
+  deriving (Generic, Show)
+
+newtype Color2 = Color2 Color
+
+instance TextShow Color2 where
+    showbPrec p (Color2 c) = genericShowbPrec p c
+
+colorShowt :: Color -> T.Text
+colorShowt c = case c of
+  Red    -> T.pack "Red"
+  Green  -> T.pack "Green"
+  Blue   -> T.pack "Blue"
+  Orange -> T.pack "Orange"
+  Violet -> T.pack "Violet"
+
+-------------------------------------------------------------------------------
+
 $(deriveTextShow ''BinTree2)
+$(deriveTextShow ''Color)