+++ tests/Instances/Data/List/NonEmpty.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP             #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
@@ -15,15 +16,17 @@ module Instances.Data.List.NonEmpty () where
 
 import Data.List.NonEmpty (NonEmpty(..))
 
+import Text.Show.Deriving (deriveShow1)
+
+#if !(MIN_VERSION_QuickCheck(2,9,0))
 import Prelude ()
 import Prelude.Compat
 
 import Test.QuickCheck (Arbitrary(..))
 
-import Text.Show.Deriving (deriveShow1)
-
 instance Arbitrary a => Arbitrary (NonEmpty a) where
     arbitrary = (:|) <$> arbitrary <*> arbitrary
+#endif
 
 -- TODO: Replace this with a non-orphan instance
 $(deriveShow1 ''NonEmpty)