+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# next
+* Add benchmark group that tests the performance of using `String` `show` followed by `Text.pack`
+
 ### 3.2.2
 * Added benchmarks