+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # next
 * Refactored the internals of `TextShow.Generic` to avoid the use of proxies.
 * Made benchmark suite more comprehensive, including benchmarks for showing an enumeration type
+* Allow building with `QuickCheck-2.9`
 * Fix GHC HEAD build
 
 ### 3.2.2